Xilpm: versal: server: Resolve misra_c_2012_rule_10_7_violation
MISRA:
- Rule 10.7: If a composite expression is used as one operand of
	     an operator in which the usual arithmetic conversions
	     are performed then the other operand shall not have
	     wider essential type.

To resolve this MISRA violations, do explicit typecast as per requirements.


Acked-for-series: Jolly Shah <jolly.shah@xilinx.com>

+++ lib/sw_services/xilpm/src/versal/server/xpm_bisr.c
@@ -690,7 +690,7 @@ static XStatus XPmBisr_RepairME(u32 EfuseTagAddr, u32 TagId,u32 TagSize,u32 TagO
 	/* Compilation warning fix */
 	(void)TagId;
 
-	BaseAddr = (u64)VIVADO_ME_BASEADDR + (TagOptional << ME_BISR_EFUSE_OFFSET_SHIFT);
+	BaseAddr = (u64)VIVADO_ME_BASEADDR + (u64)(TagOptional << ME_BISR_EFUSE_OFFSET_SHIFT);
 	BisrDataDestAddr = BaseAddr + ME_BISR_FIXED_OFFSET;
 
 	/* Copy repair data */
